Research cluster

+++ tetracorder/libraries_tetracorder.sh
@@ -14,29 +14,39 @@ echo $filebase_name
 echo $unmixing_filebase_name
 
 UNMIX=false
+TETRACORDER=false
 
+POSITIONAL_ARGS=()
+
+# Combine into ONE loop to catch all flags
 while [[ $# -gt 0 ]]; do
   case $1 in
     --unmix)
       UNMIX=true
-      shift # Move to the next argument
+      shift
+      ;;
+    --tetracorder)
+      TETRACORDER=true
+      shift
       ;;
     *)
-      # This handles positional arguments or unknown flags
-      POSITIONAL_ARGS+=("$1") 
+      POSITIONAL_ARGS+=("$1") # Save non-flag arguments
       shift
       ;;
   esac
 done
 
+# Optional: Restore positional arguments if needed for the rest of the script
+set -- "${POSITIONAL_ARGS[@]}"
+
 if [ "$UNMIX" = true ]; then
     echo "Process: Unmixing enabled."
 
     emc_out_directory=${out_base}/emc2/
-    if [ -d "$emc_out_directory" ]; then
-        echo "Directory '$emc_out_directory' exists. Removing..."
-        rm -rf "$emc_out_directory"
-    fi
+    #if [ -d "$emc_out_directory" ]; then
+    #    echo "Directory '$emc_out_directory' exists. Removing..."
+    #    rm -rf "$emc_out_directory"
+    #fi
 
     echo "Created emc2 dir: ${emc_out_directory}"
     mkdir -p ${emc_out_directory}
@@ -47,27 +57,35 @@ else
   echo "Unmixing disabled!"
 fi
 
-# run tetracorder
-tetracorder_out_directory=${out_base}/tetracorder_${filebase_name}/
 
-if [ -d "$tetracorder_out_directory" ]; then
-    echo "Directory '$tetracorder_out_directory' exists. Removing..."
-    rm -rf "$tetracorder_out_directory"
+if [ "$TETRACORDER" = true ]; then
+    echo "Process: Tetracorder enabled."
+
+    # run tetracorder
+    tetracorder_out_directory=${out_base}/${filebase_name}_${unmixing_filebase_name}/
+
+    if [ -d "$tetracorder_out_directory" ]; then
+        echo "Directory '$tetracorder_out_directory' exists. Removing..."
+        rm -rf "$tetracorder_out_directory"
+    fi
+
+    mkdir -p ${tetracorder_out_directory}
+    echo "Created tetracorder dir: ${tetracorder_out_directory}"
+
+    # augment rfl data and run tetracorder
+    python ./utils/augment_file.py ${rfl_img} ${tetracorder_out_directory} --augment # augment rfl file
+    ./tetracorder/tetracorder.sh "${tetracorder_out_directory}/${filebase_name}_aug" ${tetracorder_out_directory} emit --delete_tc_output 
+
+    # deaugment data in tetracorder output director
+    cp ${rfl_img} ${tetracorder_out_directory}
+    cp ${rfl_img}.hdr ${tetracorder_out_directory}
+    python ./utils/augment_file.py ${tetracorder_out_directory}/${filebase_name} ${tetracorder_out_directory} --deaugment
+    rm ${tetracorder_out_directory}/${filebase_name}
+    #${tetracorder_out_directory}/${filebase_name}.hdr
+
+else
+  echo "Tetracorder disabled!"
 fi
 
-mkdir -p ${tetracorder_out_directory}
-echo "Created tetracorder dir: ${tetracorder_out_directory}"
-
-# augment rfl data and run tetracorder
-python ./utils/augment_file.py ${rfl_img} ${tetracorder_out_directory} --augment # augment rfl file
-./tetracorder/tetracorder.sh "${tetracorder_out_directory}/${filebase_name}_augmented" ${tetracorder_out_directory} --delete_tc_output 
-
-# deaugment data in tetracorder output director
-cp ${rfl_img} ${tetracorder_out_directory}
-cp ${rfl_img}.hdr ${tetracorder_out_directory}
-python ./utils/augment_file.py ${tetracorder_out_directory}/${filebase_name} ${tetracorder_out_directory} --deaugment
-rm ${tetracorder_out_directory}/${filebase_name}
-${tetracorder_out_directory}/${filebase_name}.hdr
-
 # push data to drive
-/store/shared/rclone/bin/rclone copy ${out_base}/ "cdrive:${out_base#./}" -P 
+#/store/shared/rclone/bin/rclone copy ${out_base}/ "cdrive:${out_base#./}" -P
